requirements.txt
httpx
fastapi==0.135.0
uvicorn==0.41.0
pymongo==4.16.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic_core==2.41.5
dnspython==2.8.0
python-multipart==0.0.22
email-validator==2.2.0
motor==3.7.0
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
